% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInput.R
\name{allCompartments}
\alias{allCompartments}
\title{Gathering all compartments into a model}
\usage{
allCompartments(modelName, transmissionRate, ...)
}
\arguments{
\item{modelName}{name of model}

\item{transmissionRate}{transmission rate, use to compute force of infection}

\item{...}{all compartments objects created by createCompartment()}
}
\description{
Gathering all compartments into a model
}
