% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInput.R
\name{allModels}
\alias{allModels}
\title{Full JSON input format}
\usage{
allModels(
  daysFollowUp,
  errorTolerance = 0.01,
  timeStep = 1,
  modelStructure,
  infectiousComps,
  contactAssumption,
  contacts,
  models
)
}
\arguments{
\item{daysFollowUp}{number of time step to be simulated}

\item{errorTolerance}{if the cumulative probability exceed 1 - errorTolerance
it will be rounded to 1.0, default = 0.01}

\item{timeStep}{how much the time interval should be binned, default = 1 (not binning); 
adjust this to derive continuous-time model}

\item{modelStructure}{define model structure, ex: "S -> I", "I -> R"}

\item{infectiousComps}{name of the infectious compartment, ex: "I"}

\item{contactAssumption}{to be deprecated soon}

\item{contacts}{created by createContact()}

\item{models}{vector contains all model objects}
}
\description{
Generate a full JSON input format
}
