% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{constant}
\alias{constant}
\title{Define a fixed number of individuals of the left compartment transit to the
right compartment at every time step}
\usage{
constant(x)
}
\arguments{
\item{x}{number of individuals who move from one compartment to another}
}
\description{
Define a fixed number of individuals of the left compartment transit to the
right compartment at every time step
}
