% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInput.R
\name{createCompartment}
\alias{createCompartment}
\title{Define compartment}
\usage{
createCompartment(name, distribution, initialValue)
}
\arguments{
\item{name}{name of this compartment, ex: "S", "I", "R"}

\item{distribution}{distribution of this compartment, defined by createDistribution()}

\item{initialValue}{number of people in this compartment at day 0}
}
\description{
Define compartment
}
