% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInput.R
\name{createContact}
\alias{createContact}
\title{Create contact manually}
\usage{
createContact(contactType, contactClasses, contactRates)
}
\arguments{
\item{contactType}{type of the contact, example "age" for age group 0-14, 15-64...}

\item{contactClasses}{a vector contains all categories of this contact}

\item{contactRates}{contact rates between pair of categories}
}
\description{
Create contact manually
}
