% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInput.R
\name{createDistribution}
\alias{createDistribution}
\title{Define distribution}
\usage{
createDistribution(name, ...)
}
\arguments{
\item{name}{name of the distribution, could be "exponential", "gamma", "weibull",
"custom", or "transitionProb"}
}
\description{
Define distribution
}
\details{
parameters for each distribution type are:
\itemize{
 \item{"exponential": }{rate}
 \item{"gamma": }{scale, shape}
 \item{"weibull": }{scale, shape}
 \item{"custom": }{a vector of percentages}
 \item{"transitionProb": }{a constant}
}
}
