% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{d_exponential}
\alias{d_exponential}
\title{Discrete exponential distribution}
\usage{
d_exponential(rate, dist_init = FALSE)
}
\arguments{
\item{rate}{rate parameter of an exponential distribution}

\item{dist_init}{whether to distribute initial value across subcompartments following this distribution. (default to FALSE, meaning init value is always in the first compartment)}
}
\value{
a Distribution object for simulator
}
\description{
Discrete exponential distribution
}
\examples{
transitions <- list("I -> D" = d_exponential(0.3))


}
