% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{d_gamma}
\alias{d_gamma}
\title{Discrete gamma distribution}
\usage{
d_gamma(rate, shape, dist_init = FALSE)
}
\arguments{
\item{rate}{rate parameter of a gamma distribution}

\item{shape}{shape parameter of a gamma distribution}

\item{dist_init}{whether to distribute initial value across subcompartments following this distribution.}
}
\value{
a Distribution object for simulator
}
\description{
Discrete gamma distribution
}
\examples{
transitions <- list("S -> I" = d_gamma(rate = 1, shape = 5))
}
