% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{d_lognormal}
\alias{d_lognormal}
\title{Discrete log-normal distribution}
\usage{
d_lognormal(mu, sigma, dist_init = FALSE)
}
\arguments{
\item{mu}{location parameter or the ln mean}

\item{sigma}{scale parameter or ln standard deviation}

\item{dist_init}{whether to distribute initial value across subcompartments following this distribution. (default to FALSE, meaning init value is always in the first compartment)}
}
\value{
a Distribution object for simulator
}
\description{
Discrete log-normal distribution
}
\examples{
transitions <- list("I -> D" = d_lognormal(3, 0.6))
transitions <- denim_dsl({I -> D = d_lognormal(3, 0.6)})
}
