% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denim.R
\docType{package}
\name{denim-package}
\alias{denim}
\alias{denim-package}
\title{denim}
\description{
Simulate deterministic model
}
\details{
Imports
}
\seealso{
Useful links:
\itemize{
  \item \url{https://drthinhong.com/denim/}
  \item \url{https://github.com/thinhong/denim}
  \item Report bugs at \url{https://github.com/thinhong/denim/issues}
}

}
\author{
\strong{Maintainer}: Anh Phan \email{anhptq@oucru.org} (\href{https://orcid.org/0009-0000-2129-435X}{ORCID})

Authors:
\itemize{
  \item Thinh Ong \email{thinhop@oucru.org} (\href{https://orcid.org/0000-0001-6772-9291}{ORCID}) [copyright holder]
  \item Marc Choisy \email{mchoisy@oucru.org} (\href{https://orcid.org/0000-0002-5187-6390}{ORCID})
}

Other contributors:
\itemize{
  \item Niels Lohman (Author of the bundled JSON C++ library) [contributor, copyright holder]
  \item Bjoern Hoehrmann \email{bjoern@hoehrmann.de} (Author of UTF-8 validation code in the bundled JSON C++ library) [contributor, copyright holder]
  \item Florian Loitsch (Author of the binary to decimal floating-point conversion code in the bundled JSON C++ library) [contributor, copyright holder]
  \item Ingo Berg (Author of the bundled MuParser C++ library) [contributor, copyright holder]
}

}
