% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{distributionToJson}
\alias{distributionToJson}
\title{Distribution object to json}
\usage{
distributionToJson(distribution)
}
\arguments{
\item{distribution}{a list with elements $name, $rate / $scale / $shape...}
}
\value{
a json object that match format {"distribution": "weibull", "scale": 2, "shape": 5}
}
\description{
Input is a distribution object of a compartment, not the full vector/list 
of distributions
}
