% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{exponential}
\alias{exponential}
\title{Discrete exponential distribution}
\usage{
exponential(rate)
}
\arguments{
\item{rate}{rate parameter of an exponential distribution}
}
\description{
Discrete exponential distribution
}
