% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeJsonFile.R
\name{getContactMatrix}
\alias{getContactMatrix}
\title{Make contact element from a contact matrix}
\usage{
getContactMatrix(contactMatrix, contactType)
}
\arguments{
\item{contactMatrix}{contact matrix dataframe, with column names and row
names are the classes}

\item{contactType}{type of the contact, example "age" for age group 0-14, 15-64...}
}
\description{
Make contact element from a contact matrix
}
