% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inputHelper}
\alias{inputHelper}
\title{Generate correct input format}
\usage{
inputHelper(transitions, contacts = NULL)
}
\arguments{
\item{transitions}{}

\item{contacts}{}
}
\value{

}
\description{
Generate correct input format
}
