% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{lognormal}
\alias{lognormal}
\title{Discrete log-normal distribution}
\usage{
lognormal(mu, sigma)
}
\arguments{
\item{mu}{location parameter or the ln mean}

\item{sigma}{scale parameter or ln standard deviation}
}
\description{
Discrete log-normal distribution
}
