% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{mathexpr}
\alias{mathexpr}
\title{Mathematical expression}
\usage{
mathexpr(expression)
}
\arguments{
\item{expression}{math expression, can be put inside "" or not,
e.g \code{mathexpr(beta * S * I / N)} or \code{mathexpr("beta * S * I / N")}
are both acceptable}
}
\description{
User-defined mathematical expression. The expression will be processed by
\code{muparser} library which offers a wide variety of operators. Kindly visit
\href{https://beltoforion.de/en/muparser/features.php}{muparser website}
to see full list of available operators.
}
