% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{multinomial}
\alias{multinomial}
\title{Multinomial}
\usage{
multinomial(...)
}
\arguments{
\item{...}{a vector of probabilities. Vector is automatically rescaled to sum to 1.}
}
\value{
a Distribution object for simulator
}
\description{
Define a set of probabilities of transition from one compartment to multiple
compartments

\if{html}{\out{<div class="sourceCode">}}\preformatted{"I -> R" = d_gamma(1/3, 2),
"I -> D" = d_lognormal(2, 0.5)
}\if{html}{\out{</div>}}

is equal to

\if{html}{\out{<div class="sourceCode">}}\preformatted{"0.5 * I -> R" = d_gamma(1/3, 2),
"0.5 * I -> D" = d_lognormal(2, 0.5)
}\if{html}{\out{</div>}}
}
