% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{nonparametric}
\alias{nonparametric}
\title{Nonparametric}
\usage{
nonparametric(...)
}
\arguments{
\item{...}{a vector of values}
}
\description{
Convert a vector of frequencies, percentages... into a distribution
}
