% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{plot.denim}
\alias{plot.denim}
\title{Overloaded plot function for denim object}
\usage{
\method{plot}{denim}(x, ..., color_palette = NULL)
}
\arguments{
\item{x}{\itemize{
\item output of denim::sim function
}}

\item{...}{\itemize{
\item additional parameter for \code{plot()} function
}}

\item{color_palette}{\itemize{
\item a palette name from the colorspace package. You can view available palettes with \code{colorspace::hcl_palettes("qualitative", plot = TRUE)}.
}}
}
\description{
Overloaded plot function for denim object
}
