% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInput.R
\name{setInfectiousComps}
\alias{setInfectiousComps}
\title{Set name of infectious compartment}
\usage{
setInfectiousComps(...)
}
\arguments{
\item{...}{name of the infectious compartment}
}
\description{
Set name of infectious compartment
}
