% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{transprob}
\alias{transprob}
\title{Transition probability}
\usage{
transprob(x)
}
\arguments{
\item{x}{a float number between 0 to 1}
}
\description{
A fixed percentage of the left compartment transit to the right compartment
at every time step
}
