% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{weibull}
\alias{weibull}
\title{Discrete Weibull distribution}
\usage{
weibull(scale, shape)
}
\arguments{
\item{scale}{scale parameter of a Weibull distribution}

\item{shape}{shape parameter of a Weibull distribution}
}
\description{
Discrete Weibull distribution
}
