% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{constant}
\alias{constant}
\title{Constant}
\usage{
constant(x)
}
\arguments{
\item{x}{a constant number}
}
\description{
Define a fixed number of individuals of the left compartment transit to the
right compartment at every time step
}
