% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{d_gamma}
\alias{d_gamma}
\title{Discrete gamma distribution}
\usage{
d_gamma(rate, shape, dist_init = FALSE)
}
\arguments{
\item{rate}{rate parameter of a gamma distribution}

\item{shape}{shape parameter of a gamma distribution}

\item{dist_init}{whether to distribute initial value across subcompartments following this distribution.}
}
\value{
a Transition object for simulator
}
\description{
Discrete gamma distribution
}
\examples{
transitions <- list("S -> I" = d_gamma(rate = 1, shape = 5))
transitions_dsl <- denim_dsl({S -> I = d_gamma(rate = 1, shape = 5)})
# define model parameters as distributional parameters
transitions_dsl <- denim_dsl({S -> I = d_gamma(rate = i_rate, shape = i_shape)})
}
