% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{d_gamma}
\alias{d_gamma}
\title{Discrete gamma distribution}
\usage{
d_gamma(scale, shape)
}
\arguments{
\item{scale}{scale parameter of a gamma distribution}

\item{shape}{shape parameter of a gamma distribution}
}
\description{
Discrete gamma distribution
}
