% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{d_lognormal}
\alias{d_lognormal}
\title{Discrete log-normal distribution}
\usage{
d_lognormal(mu, sigma)
}
\arguments{
\item{mu}{location parameter or the ln mean}

\item{sigma}{scale parameter or ln standard deviation}
}
\description{
Discrete log-normal distribution
}
\examples{
transitions <- list("I -> D" = d_lognormal(3, 0.6))
}
