% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{d_weibull}
\alias{d_weibull}
\title{Discrete Weibull distribution}
\usage{
d_weibull(scale, shape)
}
\arguments{
\item{scale}{scale parameter of a Weibull distribution}

\item{shape}{shape parameter of a Weibull distribution}
}
\description{
Discrete Weibull distribution
}
