% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denim.R
\docType{package}
\name{denim-package}
\alias{denim}
\alias{denim-package}
\title{denim}
\description{
Simulate deterministic discrete time model
}
\details{
Imports
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thinhong.github.io/denim}
  \item \url{https://github.com/thinhong/denim}
  \item Report bugs at \url{https://github.com/thinhong/denim/issues}
}

}
\author{
\strong{Maintainer}: Thinh Ong \email{thinhop@oucru.org} (\href{https://orcid.org/0000-0001-6772-9291}{ORCID})

Authors:
\itemize{
  \item Marc Choisy \email{mchoisy@oucru.org} (\href{https://orcid.org/0000-0002-5187-6390}{ORCID})
}

}
