% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mathexpr}
\alias{mathexpr}
\title{Mathematical expression}
\usage{
mathexpr(expr)
}
\arguments{
\item{expr}{User defined mathematial expression. he expression will be processed by
muparser library which offers a wide variety of operators. Visit
muparser website (https://beltoforion.de/en/muparser/features.php)
to see full list of available operators.}
}
\value{
a Distribution object for simulator
}
\description{
Mathematical expression
}
\examples{
transitions <- list("S->I"=mathexpr("beta*S/N"))
# definition for parameters in the expression required
params <- c(N = 1000, beta = 0.3)
}
