% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mathexpr}
\alias{mathexpr}
\title{Mathematical expression}
\usage{
mathexpr(expr)
}
\arguments{
\item{exp}{math expression, a string}
}
\description{
User-defined mathematical expression. The expression will be processed by
\code{muparser} library which offers a wide variety of operators. Kindly visit
\href{https://beltoforion.de/en/muparser/features.php}{muparser website}
to see full list of available operators.
}
