% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{multinomial}
\alias{multinomial}
\title{Multinomial}
\usage{
multinomial(...)
}
\arguments{
\item{...}{a vector of probabilities, must add up to 1}
}
\description{
Define a set of probabilities of transition from one compartment to multiple
compartments\preformatted{"I -> R, D" = multinomial(0.9, 0.1),
"I -> R" = gamma(3, 2),
"I -> D" = lognormal(2, 0.5)
}

is equal to\preformatted{"0.9 * I -> R" = gamma(3, 2),
"0.1 * I -> D" = lognormal(2, 0.5)
}
}
