% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{nonparametric}
\alias{nonparametric}
\title{Nonparametric distribution}
\usage{
nonparametric(..., dist_init = FALSE)
}
\arguments{
\item{...}{a vector of values}
}
\value{
a Distribution object for simulator
}
\description{
Convert a vector of frequencies, percentages... into a distribution
}
\examples{
transitions <- list("S->I"=nonparametric(0.1, 0.2, 0.5, 0.2))
}
